Extract shared scraper utilities
